+++ MDB2/Driver/Manager/oci8.php
@@ -67,7 +67,10 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function createDatabase($name)
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
 
         $username = $db->options['database_name_prefix'].$name;
         $password = $db->dsn['password'] ? $db->dsn['password'] : $name;
@@ -107,7 +110,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function dropDatabase($name)
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $username = $db->options['database_name_prefix'].$name;
         return $db->standaloneQuery('DROP USER '.$username.' CASCADE');
     }
@@ -203,7 +210,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function alterTable($name, $changes, $check)
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         foreach ($changes as $change_name => $change) {
             switch ($change_name) {
             case 'added_fields':
@@ -304,7 +315,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function listDatabases()
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         if ($db->options['database_name_prefix']) {
             $query = "SELECT SUBSTR(username, "
                 .(strlen($db->options['database_name_prefix'])+1)
@@ -338,8 +353,12 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function listUsers()
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
-        $query = "SELECT username FROM sys.all_users";
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
+        $query = 'SELECT username FROM sys.all_users';
         $users = $db->queryCol($query);
         if (PEAR::isError($users)) {
             return $users;
@@ -362,8 +381,12 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function listViews()
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
-        $query = "SELECT view_name FROM sys.user_views";
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
+        $query = 'SELECT view_name FROM sys.user_views';
         $views = $db->queryCol($query);
         if (PEAR::isError($views)) {
             return $views;
@@ -387,7 +410,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function listFunctions()
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $query = "SELECT name FROM sys.user_source WHERE line = 1 AND type = 'FUNCTION'";
         $functions = $db->queryCol($query);
         if (PEAR::isError($functions)) {
@@ -412,7 +439,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      **/
     function listTables()
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $query = 'SELECT table_name FROM sys.user_tables';
         return $db->queryCol($query);
     }
@@ -429,7 +460,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function listTableFields($table)
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $table = strtoupper($table);
         $query = "SELECT column_name FROM user_tab_columns WHERE table_name='$table' ORDER BY column_id";
         $fields = $db->queryCol($query);
@@ -451,14 +486,16 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      * @param object $db database object that is extended by this class
      * @param string $seq_name name of the sequence to be created
      * @param string $start start value of the sequence; default is 1
-     * @param boolean   $auto_increment if the seq should be auto inc or not; default is false
-     * @param string    $field name of the field that's being turned into auto increment
      * @return mixed MDB2_OK on success, a MDB2 error on failure
      * @access public
      */
-    function createSequence($seq_name, $start = 1, $auto_increment = false, $field = '')
+    function createSequence($seq_name, $start = 1)
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $sequence_name = $db->getSequenceName($seq_name);
         return $db->query("CREATE SEQUENCE $sequence_name START WITH $start INCREMENT BY 1".
             ($start < 1 ? " MINVALUE $start" : ''));
@@ -477,7 +514,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function dropSequence($seq_name)
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $sequence_name = $db->getSequenceName($seq_name);
         return $db->query("DROP SEQUENCE $sequence_name");
     }
@@ -493,7 +534,11 @@ class MDB2_Driver_Manager_oci8 extends MDB2_Driver_Manager_Common
      */
     function listSequences()
     {
-        $db =& $GLOBALS['_MDB2_databases'][$this->db_index];
+        $db =& $this->getDBInstance();
+        if (PEAR::isError($db)) {
+            return $db;
+        }
+
         $query = "SELECT sequence_name FROM sys.user_sequences";
         $table_names = $db->queryCol($query);
         if (PEAR::isError($table_names)) {
